working state
implemented saving of a quite large data structure in another project.
This doesn't work out of the box, because when pulling this module into
the project, it screwes up other Fable.Browser modules. So had to add
the Fable.Browser.Event package from nuget for it to work. Otherwise it
seems to work fine